% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_bic_clcm.R
\name{aic_bic_clcm}
\alias{aic_bic_clcm}
\title{AIC, BIC, -2LL}
\usage{
aic_bic_clcm(mod)
}
\arguments{
\item{mod}{object from the \code{clcm()} function}
}
\value{
Returns the -2LL, AIC, and BIC of the confirmatory latent class model. Allows you to compare relative model fit.
}
\description{
Compute AIC, BIC, and 2LL of the CLCM model
}
\examples{
set.seed(3112021)
# Simulate 5 Ordinal items
sim.dat <- simulate_clcm(N = 200,
                         number.timepoints = 1,
                         item.type = rep('Ordinal', 5),
                         categories.j = rep(4, 5),
                         lc.prop = list('Time_1' = c(0.5, 0.5)) )

# Fit a model with 5 nominal items to the data
mod1 <- clcm(dat = sim.dat$dat,
            item.type = rep('Nominal', 5),
            item.names = sim.dat$item.names,
            Q = sim.dat$Q,
            verbose = FALSE)

# Fit a second model with 5 ordinal items to the data             
mod2 <- clcm(dat = sim.dat$dat,
            item.type = rep('Ordinal', 5),
            item.names = sim.dat$item.names,
            Q = sim.dat$Q,
            verbose = FALSE)       

# Compare the two model fits:
unlist(aic_bic_clcm(mod1))
unlist(aic_bic_clcm(mod2))
}
