% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C2_clcm.R
\name{C2_clcm}
\alias{C2_clcm}
\title{C2 Fit Statistic}
\usage{
C2_clcm(mod, verbose = T)
}
\arguments{
\item{mod}{estimated model object from \code{clcm()} function}

\item{verbose}{logical; print the C2 statistic and associated p-value? Default is TRUE, set to FALSE for simulations.}
}
\value{
Returns the C2 fit statistic, associated p-value (test of exact fit,
where null hypothesis is the model fits the data exactly), RMSEA value,
sample size N, degrees of freedom, observed probabilities, and model-implied probabilities.
The latter two can be used to assess the fit visually.
}
\description{
Compute C2 Test Statistic for the CLCM
}
\examples{
\dontrun{
set.seed(3112021)
sim.dat <- simulate_clcm(N=200,
                          number.timepoints = 1,
                          item.type = rep('Ordinal', 5),
                          categories.j = rep(4, 5),
                          lc.prop = list('Time_1' = c(0.5, 0.5)) )

mod <- clcm(dat = sim.dat$dat,
            item.type = sim.dat$item.type,
            item.names = sim.dat$item.names,
            Q = sim.dat$Q)

mod.fit <- C2_clcm(mod)

 }
}
