% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLCM-package.R
\docType{package}
\name{CLCM-package}
\alias{CLCM}
\alias{CLCM-package}
\title{CLCM: Estimate Confirmatory Latent Class Models}
\description{
Estimate confirmatory latent class models for a variety of item response types that are encountered in the clinical field. One or two timepoints are supported. Latent regression estimation can be performed, allowing for comparisons of longitudinal latent class assignments (e.g., treatment success/failure) across observed groups (e.g., treatment arms in clinical trials). Fit statistics (C2, AIC/BIC) are available as well.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CJangelo/CLCM}
  \item \url{https://cjangelo.github.io/CLCM/}
  \item Report bugs at \url{https://github.com/CJangelo/CLCM/issues}
}

}
\author{
\strong{Maintainer}: Charlie Iaconangelo \email{charles.iaconangelo@gmail.com}

}
\keyword{internal}
