% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_loglike.R
\name{compute_loglike}
\alias{compute_loglike}
\title{Loglikelihood}
\usage{
compute_loglike(X, item.type, param, eta, categories.j)
}
\arguments{
\item{X}{matrix of item responses}

\item{item.type}{character vector of item types}

\item{eta}{matrix operationalizing the condensation rules}

\item{categories.j}{numeric vector of number of categories per item}
}
\value{
matrix of N by 2^K
}
\description{
Compute the loglikelihood of a CLCM
}
