% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_post.R
\name{compute_post}
\alias{compute_post}
\title{Posterior distribution}
\usage{
compute_post(X, item.type, param, lprior, eta, categories.j)
}
\arguments{
\item{X}{matrix of item responses}

\item{item.type}{character vector of item types}

\item{param}{list of item parameters}

\item{lprior}{N by 2^K matrix of log of prior distribution}

\item{eta}{matrix operationalizing the condensation rules}

\item{categories.j}{numeric vector of number of categories per item}
}
\value{
matrix of N by 2^K
}
\description{
Compute the posterior distribution of a CLCM
}
