% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_bic_clcm.R
\name{aic_bic_clcm}
\alias{aic_bic_clcm}
\title{AIC, BIC, -2LL}
\usage{
aic_bic_clcm(mod)
}
\arguments{
\item{mod}{object from the \code{clcm()} function}
}
\value{
Returns the -2LL, AIC, and BIC of the confirmatory latent class model.
}
\description{
Compute AIC, BIC, and 2LL of the CLCM model
}
\examples{
\dontrun{
set.seed(3112021)
sim.dat <- simulate_clcm(N=1000,
                         number.timepoints = 1,
                         item.type = rep('Ordinal', 5),
                         categories.j = rep(4, 5),
                         lc.prop = list('Time_1' = c(0.5, 0.5)) )


mod1 <- clcm(dat = sim.dat$dat,
            item.type = rep('Nominal', 5),
            item.names = sim.dat$item.names,
            Q = sim.dat$Q)


mod2 <- clcm(dat = sim.dat$dat,
           item.type = rep('Ordinal', 5),
           item.names = sim.dat$item.names,
           Q = sim.dat$Q)


mod.fit1 <- aic_bic_clcm(mod1)
mod.fit2 <- aic_bic_clcm(mod2)


unlist(mod.fit1)
unlist(mod.fit2)
}
}
