% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clcm_2stage.R
\name{clcm_2stage}
\alias{clcm_2stage}
\title{Two-Stage Estimation of the CLCM}
\usage{
clcm_2stage(dat, clcm.mod, lprior.t1)
}
\arguments{
\item{dat}{dataframe containing the new item responses you want to fit the
model to.}

\item{clcm.mod}{a CLCM model that you already fit (Stage 1). This will
have the item parameter estimates, item names, item type, and Q-matrix.}

\item{lprior}{this is the log-prior for timepoint 2. N by L matrix.}
}
\description{
Treat item parameters as fixed to fit a longitudinal CLCM. The model
can have any number of timepoints. The only limitation is the estimation
of the posteriors - item parameters are fixed.
The CLCM R package requires that count items start at 0
}
