% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lprior.R
\name{compute_lprior}
\alias{compute_lprior}
\title{Prior Distribution}
\usage{
compute_lprior(post, K, Z = NULL, type = NULL, reg.formula = NULL)
}
\arguments{
\item{post}{posterior distribution}

\item{Z}{data frame or matrix containing covariates}

\item{type}{the type of prior distribution - this is selected automatically in the clcm() function
depending on what is passed}

\item{reg.formula}{if a latent regression is estimated, pass the model specification as a character here}
}
\value{
matrix of N by 2^K
}
\description{
Compute the log of the prior distribution for use in the EM algorithm
}
