% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starts_with_zero.R
\name{start_with_zero}
\alias{start_with_zero}
\title{Count Items Start with Zero}
\usage{
start_with_zero(dat, items)
}
\arguments{
\item{dat}{dataframe containing the item responses}

\item{items}{vector of item names. These should be the count items that need
to start at zero in order for \code{clcm} function to estimate the model.}
}
\description{
The CLCM R package requires that count items start at 0
}
